Make the testcase a bit more robust
Wait for longer. Sometimes, the process is dead after the first ctrl+c,
and we must not send a signal in that case. Add some debugging output,
which should also ensure the stderr pipe is not filling up.

+++ tests/server/test_startup.py
@@ -289,11 +289,11 @@ async def test_libertem_server_cli_startup():
     # total deadline, basically how long it takes to import all the dependencies
     # and start the web API
     # (no executor startup is done here)
-    deadline = time.monotonic() + 5
+    deadline = time.monotonic() + 15
     while True:
         if time.monotonic() > deadline:
             assert False, 'timeout'
-        line = await asyncio.wait_for(p.stderr.readline(), 2)
+        line = await asyncio.wait_for(p.stderr.readline(), 5)
         if not line:  # EOF
             assert False, 'subprocess is dead'
         line = line.decode("utf8")
@@ -301,12 +301,23 @@ async def test_libertem_server_cli_startup():
         if 'LiberTEM listening on' in line:
             break
 
+    async def _debug():
+        while True:
+            line = await asyncio.wait_for(p.stderr.readline(), 5)
+            if not line:  # EOF
+                return
+            line = line.decode("utf8")
+            print('Line@_debug:', line, end='')
+
+    asyncio.ensure_future(_debug())
+
     try:
         # now, let's kill the subprocess:
         # ctrl+s twice should do the job:
         p.send_signal(signal.SIGINT)
-        await asyncio.sleep(0.1)
-        p.send_signal(signal.SIGINT)
+        await asyncio.sleep(0.5)
+        if p.returncode is None:
+            p.send_signal(signal.SIGINT)
 
         # wait for the process to stop, but max. 1 second:
         ret = await asyncio.wait_for(p.wait(), 1)
